avocado-instrumented: move logging code to runner
